requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
umap-learn
idx2numpy
jupyter